Fixed map pan not working by adding DocType
The page is missing a doctype, which sets document.compatMode to
"BackComapt". When this happens, Dojo changes how it conveys which mouse
button is clicked. With the wrong compatMode ("BackCompat" while running
in a modern browser), panning breaks because our API is testing for
mouse
clicks thinking it's in an old browser but it's really in a modern
browser. Take a look at the dojo/mouse tests for "dom-quirks" (which is
triggered when document.compatMode is "BackCompat") to see the
underlying
issue:  
http://trac.dojotoolkit.org/browser/dojo/branches/1.7/dojo/mouse.js#L36